+++ src/backend/joanie/core/api/admin/__init__.py
@@ -594,6 +594,8 @@ class NestedCourseProductRelationOrderGroupViewSet(
     permission_classes = [permissions.IsAdminUser & permissions.DjangoModelPermissions]
     serializer_classes = {
         "create": serializers.AdminOrderGroupCreateSerializer,
+        "update": serializers.AdminOrderGroupUpdateSerializer,
+        "partial_update": serializers.AdminOrderGroupUpdateSerializer,
     }
     default_serializer_class = serializers.AdminOrderGroupSerializer
     queryset = models.OrderGroup.objects.all().select_related("course_product_relation")
@@ -645,7 +647,6 @@ class OrderViewSet(
         "contract__definition",
         "certificate",
         "certificate__certificate_definition",
-        "order_group",
         "credit_card",
     )
 
@@ -781,3 +782,18 @@ class OrganizationAddressViewSet(
             ) from error
 
         return super().destroy(request, *args, **kwargs)
+
+
+class DiscountViewSet(
+    mixins.CreateModelMixin,
+    mixins.UpdateModelMixin,
+    mixins.DestroyModelMixin,
+    mixins.ListModelMixin,
+    viewsets.GenericViewSet,
+):
+    """Admin Discount Viewset"""
+
+    authentication_classes = [SessionAuthenticationWithAuthenticateHeader]
+    permission_classes = [permissions.IsAdminUser & permissions.DjangoModelPermissions]
+    serializer_class = serializers.AdminDiscountSerializer
+    queryset = models.Discount.objects.all()