+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to
 
 ### Added
 
+- Add `Discount` relation to `OrderGroup` model
 - Add `start` and `end` datetime fields on order group model
 - Add Sarbacane newsletter client